+++ tree-sitter-java/TreeSitter/Java/AST.hs
@@ -3,7 +3,8 @@ module TreeSitter.Java.AST
 ( module TreeSitter.Java.AST
 ) where
 
-import TreeSitter.GenerateSyntax
-import qualified TreeSitter.Java as Grammar
+import           TreeSitter.GenerateSyntax
+import           TreeSitter.Java (tree_sitter_java)
+import qualified TreeSitter.Java.Grammar as Grammar
 
-astDeclarationsForLanguage Grammar.tree_sitter_java "../../vendor/tree-sitter-java/src/node-types.json"
+astDeclarationsForLanguage tree_sitter_java "../../vendor/tree-sitter-java/src/node-types.json"