+++ tree-sitter-java/ChangeLog.md
@@ -1,3 +1,4 @@
 # v0.3.0.0
 
 * Adds AST datatypes in `TreeSitter.Java.AST` which you can use with `TreeSitter.Unmarshal`.
+* Moves the `Grammar` datatype into `TreeSitter.Java.Grammar`.